Initial PR feedback from Simon
- Add stub for `mfc0` to `MipsArch`
- Rename `DoubleToIntIrPattern` to `FloatishToIntIrPattern` for PPC
    - Make it more explictly handle either f32/f64
- Add `Type.sintish()` & `Type.uintish()`; use them for PPC float conv
- Remove "TODO" for handling PPC float-to-int

+++ tests/end_to_end/float-conversions/mwcc-o4p-out.c
@@ -7,5 +7,5 @@ void test(void) {
     u = __cvt_fp2unsigned(flt);
     u = __cvt_fp2unsigned((f32) dbl);
     dbl = (f64) u;
-    flt = (f32) (f64) u;
+    flt = (f32) u;
 }
